core: use ping2, confirmConnectivity verbs
For cluster compatibility >= 4.2 Host.ping2 verb replaces Host.ping
verb when engine polls the host to ascertain that vdsm is reachable.
A new verb Host.confirmConnectivity is introduced when engine polls
the host to allow vdsm to ascertain that the engine is reachable,
e.g. following network update\setup.
This patch adds support in vdsbroker for usage of the new verbs by
engine.

Change-Id: I16643614ae8b9160fa531e2777279130a25c75fd
Bug-Url: https://bugzilla.redhat.com/1525637
Signed-off-by: Eitan Raviv <eraviv@redhat.com>