engine: use ping2, confirmConnectivity verbs
For cluster compatibility >= 4.2 Host.ping2 verb replaces Host.ping
verb when engine polls the host to ascertain that vdsm is reachable.
A new verb Host.confirmConnectivity is introduced when engine polls
the host to allow vdsm to ascertain that the engine is reachable,
e.g. following network update\setup.
In this patch the different flows start using the new verbs per the
use case and the cluster compatibility version.

Change-Id: Ib86e0094e160a2ab2eb94117dace6fa52f719201
Bug-Url: https://bugzilla.redhat.com/1525637
Signed-off-by: Eitan Raviv <eraviv@redhat.com>