core: make LibvirtVmXmlBuilder a stand-alone utility
Previously, LibvirtVmXmlBuilder depended on VmInfoBuilderImpl to
generate the vm-info map that we used to send to VDSM. This patch
breaks that dependency by replacing all usages of that map (createInfo).

Not only that this change makes run-VM more efficient, as less code
is executed and in particular less interaction with the database, but
it also makes LibvirtVmXmlBuilder a stand-alone utility that can be
used to generate engine-XML for OVF of hosted-engine VM without
having to create the vm-info map first.

Change-Id: Ie3f5217b60a3d3034b4ba194ebc0de392c2d2b43
Bug-Url: https://bugzilla.redhat.com/1505333
Signed-off-by: Arik Hadas <ahadas@redhat.com>