webadmin: Change the lines' color on iSCSI LUNs table
Screenshot: https://imgur.com/a/0pXED

In case the LUN can be added / extended, its line should be black.
By pressing the "Add" or "Extend" buttons, the line turns to bold-black
If no operation is available for a LUN, it should be grayed out.
When removing LUNs, lines have strike line.

Change-Id: I5092982cdb651b0a6c39eb90e83e9d9005c06aac
Bug-Url: https://bugzilla.redhat.com/1532040
Bug-Url: https://bugzilla.redhat.com/1532844
Signed-off-by: Shani Leviim <sleviim@redhat.com>