core: send engine xml also when memory is restored
Restoring hibernated VMs or previewing snapshots with memory got
broken by the engine-XML changes since the engine doesn't provide
VDSM with the VM configuration anymore but merely provides the
metadata and memory dump volumes. Thus, VDSM doesn't have the
latest volumes that the VM should run with and therefore uses
the ones that appear in the metadata volume, which is wrong.

Now, the engine also sends the engine-XML so VDSM will know the
volumes that the VM should be started with and also the needed
information about the graphics devices.

Change-Id: Ie88975796a7d6f727b49d68554a704b15807ecb4
Signed-off-by: Arik Hadas <ahadas@redhat.com>