core: v2v - convert identical disk interface for KVM external provider
Importing a vm from KVM external provider always resulted in setting
the disk interface of target oVirt VM to be VIRTIO, regardless to the disk
interface as set in source VM.
Since virt-v2v is not called for KVM external provider and no changes are
done in target guest (i.e. kvm2ovirt just copies disks as is), then oVirt
target guest may end up with disks marked as VIRTIO which are different
than source VM and with device map file containing SCSI or IDE disk
interfaces (since device map file was copied from source VM as is).

This patch fix that behaviour so that disk(s) interface of target will be
the same as in source VM, with a limitation of converting SCSI, USB and
SATA interfaces to SCSI interface due to VDSM bug of sending the same
dev type for all those above 3 interfaces.
This limitation should be fixed in the future.

Change-Id: I1c9796871a2e0ab9e46a38c1c75f70cccfcd09f3
Bug-Url: https://bugzilla.redhat.com/1484199
Signed-off-by: Sharon Gratch <sgratch@redhat.com>