core: Rename VDSCommand's CreateSnapshot to CreateVolume
CreateSnapshotVDSCommand in practice calls the createVolume verb in
vdsm. While the term "CreateSnapshot" may have made sense
historically, nowadays its confusing, as evident by the reference
bugzilla. Creating a snapshot is an engine level term (as is
properly used by the CreateSnapshotCommand. The VDS commands, on the
other hand, are mere wrappers to call vdsm, and shouldn't refer to
the engine context they're used from. In this case, creating a volume
may indeed be used to create a snapshot, but may also be used when
creating an empty image structure (e.g., when cold-moving a disk), and
thus may be confusing.

This patch renames the CreateSnapshotVDSCommand class to
CreateVolumeVDSCommand. Similarly, the
CreateSnapshotVDSCommandParameters class is renamed to
CreateVolumeVDSCommandParameters.

Change-Id: I695aa12335e069fcbbdcffa52f0682bbb6b9fa8e
Bug-Url: https://bugzilla.redhat.com/1534227
Signed-off-by: Allon Mureinik <amureini@redhat.com>