restapi: register vnic profile by name - accept partial input
Remove all validations on vnic mapping input when it is passed
inside the registration_configuration block. Accept partial
user input in source and target of mappings so that best effort
matching and assignment of vnic profiles to vnics may take place
in the bll layer.

This is a backport from master.

Change-Id: I3b7b87b95684c86fd3eef4abb4833342fbc5da36
Bug-Url: https://bugzilla.redhat.com/1530814
Signed-off-by: Eitan Raviv <eraviv@redhat.com>