core: fix NPE when migrating vm disappears
In patch I681021d1 we generalized the generation of the abort-migration
audit log message so it can be reused for the case where a migrating vm
disappeared. However, the VmAnalyzer#auditVmMigrationAbort method was
not changed and therefore may produce an NPE in some cases. This patch
changes that method as well together with other statement that may fail
on NPE because of the same issue.

Change-Id: I1d130e192fa2f684ddb8bd775de470233e0b4872
Bug-Url: https://bugzilla.redhat.com/1505399
Signed-off-by: Arik Hadas <ahadas@redhat.com>