core: v2v - update non-bootable disks for non KVM external providers
In case of importing a VM from a non KVM external provider,
diskVmElement entities are initially set with a default value
and afterwards are overridden by data provided in the OVF
generated by virt-v2v.

There was a bug that prevented from updating the diskVmElement
in DB for each non-bootable disk, such that all non-bootable disks
were left with default values and the data, generated by virt-v2v in
OVF file, was ignored. That means that disk interface for those non-bootable
disk(s) was left with VIRTIO defalt value.
Only bootable disk(s)'s diskVmElement was updated in DB with OVF file values.

This is fixed in this patch.

Change-Id: I9f9553c513bfe8fab0374e09e4bb01dfd237d54b
Bug-Url: https://bugzilla.redhat.com/1484199
Signed-off-by: Sharon Gratch <sgratch@redhat.com>