core: extract ova out of transaction
The commands ImportVmFromOva and ImportVmFromExternalProvider are non-
transactive and use compensation. This causes their end-action phase
to be transactive. The fact that, unlike OVA conversion using virt-v2v,
the extraction of an OVA that was generated by oVirt is done in the
same thread that executes the end-action phase cause it to be executed
within a transaction and so if extracting the OVA file takes long
time the transaction may reach its timeout and the command will fail.

Change-Id: I4dba01ea768565aed82e79e2df54e82b294361dc
Bug-Url: https://bugzilla.redhat.com/1535521
Signed-off-by: Arik Hadas <ahadas@redhat.com>