db: Add engine-psql.sh to dbscripts
Script engine-psql.sh was part of redhat-support-plugin-rhev and it was referenced
from many KCS articles. In 4.2 redhat-support-plugin-rhev is obsoleted by rhvm-doc,
but engine-psql.sh is not part of rhvm-doc package.
This patch adds engine-psql.sh to our dbscripts directory.

The script was changed to support SCL and non-SCL PG

Also added logging to /var/log/messages if rsyslog is installed [1]

[1] https://fedoraproject.org/wiki/Changes/NoDefaultSyslog

Change-Id: I4d9737ea72df0d7e654776a1085901284a523b7f
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1525904
Signed-off-by: emesika <emesika@redhat.com>