core: Fix role mapping for vm and template registration
Fix NPE when registering a Template.
The NPE was caused when trying to import a Template with no RoleMapping.
The role mapping should not be part of the validate phase, it should
only be part of the init process.
The following patch use RoleMapping only in the init for VM and Template
and validate the entity's Roles and Users in the validation process.

Change-Id: Ia6ad2764cb6fd1d3ea6f91822a4e63710503385b
Bug-Url: https://bugzilla.redhat.com/1536966
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>