search: use distinct on cross ref
Patch [1] improved performance by omitting DISTINCT from the search
queries when it is not needed.
After this patch we can configure each query to use or not to use the
DISTICT keyword.
However, this is true only for direct search and not for cross-reference
searches (i.e. VMs that were created from a specific template)

This patch adds the DISTINCT keyword wherever it recognized a
cross-reference.

[1] https://gerrit.ovirt.org/#/c/84340/

Change-Id: I60d817708900eb15248d91eb2d149c84a98eb1cc
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1532802
Signed-off-by: emesika <emesika@redhat.com>