webadmin: avoid removing Edit button for template sub tabs
For all SubTabPanelView entities (extends from
AbstractSubTabPanelView) we needed to avoid displaying
the 'New' button since there is no meaning in that context,
but TemplateSubTabPanelView is a special case in which there
is no button to remove when displaying it (no new action),and
since the removed button was always the leftmost one, then
there was a bug of removing the 'Edit' button.

This patch avoid removing the 'Edit' button when
creating the TemplateSubTabPanelView entity.

Change-Id: I71655601247de15f2bbf89cb5fdfb327a212e7ec
Bug-Url: https://bugzilla.redhat.com/1512554
Signed-off-by: Sharon Gratch <sgratch@redhat.com>