core: Split log message by new line
In commit 46d81cc1 (core: Display async tasks when fail to move SPM to
maintenance), we added a log message to show the details of the
running asyn tasks. As there could be multiple tasks and as the info
may be a bit long, in this patch the separator is changed from comma
to new line.

Change-Id: I30087c6f7510afb31221ee2c48c2353d477970d8
Signed-off-by: Ala Hino <ahino@redhat.com>