core: pass storage domain id to addOvfStoreDiskToDomain
When storage domain attached to a data center all the
OVF_STORE disks registered and the ovf_info entity of
each disk is added to the DB accordingly.

One of the ovf_info fields is storage_domain_id which
set to 'null' - getStorageDomainId() returns null in
StorageHandlingCommandBase::addOvfStoreDiskToDomain.

Send the storage domain id when call the method instead
of using getStorageDomainId() which returns null in
StorageHandlingCommandBase::addOvfStoreDiskToDomain

Change-Id: Id8a19a4667fbc962eb8720208263f5547691dad1
Bug-Url: https://bugzilla.redhat.com/1539361
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>