core: handle replication not in progress
Currently, when an error is returned from
VmReplicateDiskFinish engine will clean up
the failed LSM leftovers.

This patch adds a special treatment for when
the cause of failure is a replication not being
in progress, in such case it is safe to remove
the target leftovers to make the next LSM attempt
simpler.

Change-Id: Ic03d2ff31b690f9b2153de941878f3b8a3a5d0bf
Bug-Url: https://bugzilla.redhat.com/1517540
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>