core: adjust ova extraction to python 2
The function that converts a number field to a python number (nti) that
we use when extracting an OVA file, was taken from python 3's tarfile.py.
This caused an error when trying to extract OVA files that contain disks
whose size is larger than 8gb, due to some differences between python 2
and python 3. This patch changes this function so it would work properly
with both versions of python and for every disk size.

Change-Id: Ice7e11763d23b341b616955d453684b6c2838fb2
Bug-Url: https://bugzilla.redhat.com/1535574
Signed-off-by: Arik Hadas <ahadas@redhat.com>