core: Fix validation and test in LiveMigrateDiskCommand
Fix the validation in LiveMigrateDiskCommand to return early upon
failure in the validation instead of continuing to run more validations.
Since the test was build with the assumption that more validations will
run after the basic ones failed it did not bother to mock the behavior
correctly and had to be fixed as well.

Change-Id: I9758a8fb0f9ad13797c18f40e98962e27c0a20fb
Signed-off-by: Tal Nisan <tnisan@redhat.com>