core: stop image transfer when there is no activity
When there's no activity with an image transfer (for any reason), the
engine keeps monitoring it forever although it may have already stopped
(and not just delayed).
Therefore, this patch stops a transfer when there's no activity with it
according to the new config value,
TransferImageInactivityTimeoutInSeconds, for any transfer direction and
client (frontend/SDK).
For upload flows - the transfer is paused as it can be resumed.
For download flows - the transfer is canceled as it cannot be resumed
anyway.

Change-Id: Iad3a3236d7bc181fb3ceac729ab5401c21455e43
Bug-Url: https://bugzilla.redhat.com/1539363
Signed-off-by: Idan Shaby <ishaby@redhat.com>