core:restart non responding hosts after quite time
When engine is starting up , it ignores any non-responding hosts for a
configurable quite time (5 minutes by default) in order to prevent
"fencing-storm".

The problem is that hosts that remain non-responding after this quite
time are not fenced automatically and the administrator should fence them
manually.

This patch addresses a RFE that asks that after the startup quite time,
the engine will check for any non-responding hosts and if it find any, it
will try to fence them.

The patch uses the PMHealthChecker service code to grab all hosts at
startup and then after the quite time to look for each host if it still
valid (it may be deleted for example) and if it has power-management
enabled flag active and is in non-responding state, an attempt to fence
the host is done.

Change-Id: I1a3e30680d9eba7c0b43e6c99ce50ba5d90f0147
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1520424
Signed-off-by: emesika <emesika@redhat.com>