core: validate vm running before setDestroyOnReboot
Currently, there is an assumption that if there's next-run configuration
then the VM is up which is wrong.

For e.g when powering-off a VM while it is updated and a lease is added,
the engine will fail to acquire lock on the VM in order to apply next
run configuration, in case the user will select to update the VM again
while the VM is down the update will fail.

Change-Id: I2fd969af2a38925fda9435e25c167d06025d68bf
Bug-Url: https://bugzilla.redhat.com/1537534
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>