core: fix hot-(un)plug xml of NICs for hosted-engine
The emulated-machine type is not set for HE VM (since it is not
started by the engine). That caused an NPE when trying to hot-
plug/unplug a NIC to a HE VM with engine XML (cluster level > 4.1).
So we'll now set the emulated machine type according to the
cluster architecture in this case.

Change-Id: Ic3a6dcdd39f861aec33b3dfc4a473441b08a5812
Bug-Url: https://bugzilla.redhat.com/1539356
Signed-off-by: Arik Hadas <ahadas@redhat.com>