core: destroy paused vms that appear on unexpected hosts
This patch changes VMs monitoring to destroy VMs that are detected
in Paused status on a host that is different than the one we expect
the VM to run on (run_on_vds). Two exceptions though:
1. When the VM is in Unknown status in the database - that is in
order to support a scenario where the VM automatically starts on
a different host (e.g., when configured with a lease) and is
immediately paused.
2. When the VM is migrating to this host.

Change-Id: I8a79af07711063d089ff3b0f0fb7b6e12def5568
Bug-Url: https://bugzilla.redhat.com/1459865
Signed-off-by: Arik Hadas <ahadas@redhat.com>