packaging: setup: Filter engine sensitive keys
engine conf has an item SENSITIVE_KEYS with a list of keys that should
not be logged. The engine has code to filter them out. Add such code
also to engine-setup.

In practice this isn't a big issue, because the important keys are
already filtered out elsewhere. The one that is not is
ENGINE_EXTERNAL_PROVIDERS_TRUST_STORE_PASSWORD, which defaults to
'changeit', and currently we do not have documentation about how to
change that.

Change-Id: I261fd8b6fe73639978189afcf5a097912504b083
Signed-off-by: Yedidyah Bar David <didi@redhat.com>