core: align hardware resource type with ovf specification
The new reader and writer of OVFs that reside in OVAs are supposed
to be aligned with the OVF specification so it would be possible
to process both OVAs that are created by oVirt and those that are
created by other systems using the same code.

Unfortunately, oVirt used to generate Monitor and Graphics devices
with resourceType values that conflict with the OVF specification.
This patch changes those resourceTypes to be aligned with the OVF
specification, while preserving backward and forward compatibility
for OVFs that are generated by oVirt.

Change-Id: I97b01da993c4d93b092cb4b9a49ca429e9ec3c0c
Bug-Url: https://bugzilla.redhat.com/1534644
Signed-off-by: Arik Hadas <ahadas@redhat.com>