engine : Vm run fails due to NPE when host statistics are missing
When a host is added, the status of the host
is set to UP before the VdsStatistics for the
host is obtained. This causes a race where the
scheduler can throw a NPE when trying to run a
VM of the host before the statistics are obtained.

The status of the host should be marked as UP
only after the VdsStatistics has been updated
for the host.

Change-Id: Ic3b442220c738352c176d86c0a7cb55d672d890e
Bug-Url: https://bugzilla.redhat.com/1514853
Signed-off-by: Ravi Nori <rnori@redhat.com>