core: Filter versions in addHostNetworksAndUpdateInterfaces
VdsBrokerObjectsBuilder#addHostNetworksAndUpdateInterfaces attempts
to assign a switch type according to the maximal cluster version the
host supports. However, if this version is newer than the maximal
cluster version supported by the engine (i.e., a new vdsm connected
to and older engine scenario), this will fail as the engine has no
configuration for such "futuristic" versions.

This patch takes the maximum version the host support that is also
supported by the engine instead, in order to avoid such failures.

Change-Id: I449f86803e82b8816a2e41e6ce4aab6a0c5e48cf
Bug-Url: https://bugzilla.redhat.com/1548819
Signed-off-by: Allon Mureinik <amureini@redhat.com>