core: Partial import for Template with missing SDs
In case part of the Template's disks' storage domains does not exists in
the data center and allow_partial flag is enabled the Template should be
registered successfully.
There is a usecase which the Host is still connected to the storage
domains although the storage domain does not exists in the engine.
In that usecase, the validation which checks if the image exist on the
storage domain is valid although we fail to register the Template after
when we find that there are missing storage domain.
This patch should fix that scenario.

Change-Id: I114f93cd9a8317e263ba22660ec7b8c8469239cd
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>