packaging: setup: postgres95: Pass locale
Pass lc_ctype, lc_collate and encoding from the old database.
These are the values that pg_upgrade checks [1]. By default, it uses the
system's locale, so will fail if system locale is different from the old
database's (which can happen either if the old database was created with
a custom locale or if system local was changed since creating it).

[1] https://git.postgresql.org/gitweb/?p=postgresql.git;a=blob;f=src/bin/pg_upgrade/check.c;h=8d4f254;hb=HEAD#l301

Change-Id: I237a7ab6b0153a5c1f317b04c13d98cce3e775eb
Bug-Url: https://bugzilla.redhat.com/1528371
Signed-off-by: Yedidyah Bar David <didi@redhat.com>