core: Auto SD selection on template version update
Adding support of auto storage domain selection feature of VM pools when
a new template version is created for the template the pool is based on.
At this moment pool VMs are removed and created again and auto storage
domain option must be respected on creation exactly as it was when the
pool was created initially.

Change-Id: I55f5a600dd1e4d4e9738aa3cb8b1796c1455265f
Bug-Url: https://bugzilla.redhat.com/1547162
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>