core: support iothreads with scsi disks for new engine xml
Disk devices with VIRTIO-SCSI interface, that are attached to VMs
with IO threads number > 0, will now be pinned to the correct io
thread in engine xml as follows:
a SCSI controller will be created per each io thread and pinned to
that io thread.
Each SCSI disk will be addresed to use one of the SCSI controllers,
using round-robin algorithm.

Change-Id: I7d5fb6e8a048986f66d2d8a15716775d18c59bdc
Bug-Url: https://bugzilla.redhat.com/1535961
Signed-off-by: Sharon Gratch <sgratch@redhat.com>
Signed-off-by: Arik Hadas <ahadas@redhat.com>