core: Reduce rate of log events during sync
The flood rate of the PROVIDER_SYNCHRONIZED_FAILED events in
the audit log and in the webadmin's Notification Drawer is
limited to a single notification per day.
The PROVIDER_SYNCHRONIZED event is not triggered anymore.
If the synchronization succeeded, no event will be logged to
audit log, and engine.log will contain the information
that the synchronization is triggered and acquired and
released a lock.

The PROVIDER_SYNCHRONIZED is not removed from AuditLogType,
because reading PROVIDER_SYNCHRONIZED from db would create
an unexpected null pointer.

Change-Id: I5a21c54741cf34b66c1bc5f908b275a5fe0936b0
Bug-Url: https://bugzilla.redhat.com/1553126
Signed-off-by: Dominik Holler <dholler@redhat.com>