engine: Update binding:host-id during external port create/update
When a port is created/updated, it's "binding:host_id"
attribute should be updated with the id of the provider
driver id (for example OVN chassis id) reported during
get_caps.

When no provider driver id is reported, the "binding:host_id"
is set to the host id as it was set previously.

Change-Id: Ia405fee3d2d87139d777f9150064d931f5b70697
Bug-Url: https://bugzilla.redhat.com/1549030
Related-To: https://gerrit.ovirt.org/#/c/87329/
Related-To:  https://gerrit.ovirt.org/#/c/87369/
Related-To:  https://gerrit.ovirt.org/#/c/88399/
Signed-off-by: Marcin Mirecki <mmirecki@redhat.com>