dbscripts: Add provider_binding_host_id table and get/update functions
This table will contain information about the ids of
network provider drivers on a host.

GetCapabilities information can contain information about
the ids of network provider plugins on a host.
This is received as the 'openstack_binding_host_ids'
property of caps.
This information must be stored, so it can be passed
to network providers.
There can be multiple plugins per host, so multiple
key/value pairs can be received.

Change-Id: Idb5c1edd0fda3a46c633ebd003537f240797076b
Bug-Url: https://bugzilla.redhat.com/1549030
Related-To: https://gerrit.ovirt.org/#/c/87329/
Related-To:  https://gerrit.ovirt.org/#/c/87369/
Signed-off-by: Marcin Mirecki <mmirecki@redhat.com>