core: Back to List<DiskImage> on snapshot preview & restore
This reverts commit 0c6e7b, which uses a Set<Guid> for disks supplied
as parameters on REST API when previewing or restoring a snapshot.

When using the REST API for previewing or restoring a snapshot, and
supplying disks as parameters, it also needed to supply their image ids.
Using a Set<Guid> for those image_ids makes it easier to maintain,
and therefore a List<DiskImage> wasn't necessary.

Although, it's not enough validating that the 'image_id' and 'disk id'
values are not empty, but also needed to check their existence on DB.
On current code, each diskImageDao.getSnapshotById() query returns a
DiskImage for the provided image_id value (if such exists), with no
referencing the disk id (it actually being ignored).
Therefore, validating that the disk id fits the image_id gets things
more complex.

As a result, this patch uses a List<DiskImage> as before the 0c6e7b
commit, and creates a new DB query, which returns a DiskImage entity
with the supplied disk id and image_id (if such exists), or fails with
404. Since it's easy to extract those values from each DiskImage, the
DiskImage's existence validation makes more sense.

Change-Id: I59075a6ca5a993e96da1a868cc41d7b9779d8242
Bug-Url: https://bugzilla.redhat.com/1503172
Bug-Url: https://bugzilla.redhat.com/1537119
Bug-Url: https://bugzilla.redhat.com/1503038
Signed-off-by: Shani Leviim <sleviim@redhat.com>