core: Remove SearchQueryParsingUtils
Commit b39cd0 introduced the SearchParsingUtils class which was
virtually identical to SearchQueryParsingUtils, but neglected to
remove the old class. It seems that this was a botched attempt to move
the class, where the addition of the new file was committed, but the
removal of the old one was not.

This patch finishes the job and removes the old, unused,
SearchQueryParsingUtils class.

Change-Id: Ib0e6007d2106e428a0acadaa7e29dbad77e9239d
Signed-off-by: Allon Mureinik <amureini@redhat.com>