host/kernel: add option to blacklist nouveau driver
When using vGPU, the nouveau driver needs to be blacklisted. In order
to avoid forcing system administrators and users to touch the
hypervisor, we can do that via the kernel line host deploy tool.

No changes in host deploy are necessary as the whole command line is
constructed in the engine.

Change-Id: Ic45ff5de25197c258c5afcf042c1fc976f13141d
Signed-off-by: Martin Polednik <mpolednik@redhat.com>