core: fixing removal of alerts/events
When you remove events all alerts are removed as well and vice versa.
The SPs involved were changed in 4.2 [1] in order to prevent
deadlock so this bug is not a regression

The reason for that was that there was a missing join between the
results in the FROM clause and the original audit_log table on the
audit_log_id key.

This patch fixes that, separating between display/removal of
alerts/events.

[1] https://gerrit.ovirt.org/#/c/85708/

Change-Id: I7bac78ddb451c00cb6c605925df65e43688f4475
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1552473
Signed-off-by: emesika <emesika@redhat.com>