packaging: setup: Use real ISO name instead of the link
We previosuly used the link to the latest version of the RHV Guest
Tools ISO in order to upload it into the ISO domain.

That caused the filename in the ISO domain to be always
rhev-tools-setup.iso

This patch modifies the upload code, to dereference the link and
use the real filename instead.

It also fixes the issue, causing the engine to be unable to
properly identify the RHV Guest Tools ISO in the ISO domain,
since it searches it by the real name.

Change-Id: I008e5cbf2f0b0c1adb8a88af522b346d221c2517
Signed-off-by: Lev Veyde <lveyde@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1552194
(cherry picked from commit 907b4f0e08d3c80d2e7121128416c079341b312a)