webadmin: map edited VM name while importing VMs
Before this patch, when switching between VMs entities that selected
for import, the VM name in the editable checkbox was overridden by the
original VM name.

This fix provides a logic to map and persist each imported VM to its
edited name.

This fix also a regression caused by commit -
576b73f15d443d406e71de7e4537e782f0e7034d

Change-Id: If18f2b299bb0e930571fc815150407d6d1f6ec09
Bug-Url: https://bugzilla.redhat.com/1520455
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>