core: delete LiveStorageMigrationHelper
LiveStorageMigrationHelper is a singleton with a single method that gets
a CommandBase object and acts upon it.
The only consumer of LiveStorageMigrationHelper is
LiveMigrateDiskCommand which uses it twice and passes itself as the
CommandBase object.
This is a classic usecase of a private method within
LiveMigrateDiskCommand, and therefore this patch moves that method to
the command and removes LiveStorageMigrationHelper completely.

Change-Id: I91eb33abf71dcf4dcbee6b226a6f2113521ddb8f
Related-To: https://bugzilla.redhat.com/1534241
Signed-off-by: Idan Shaby <ishaby@redhat.com>