core: fix LiveMigrateDiskCommand race
One of the LiveMigrateDiskCommand steps is removing the source image.
Since RemoveImage is called using runInternalActionWithTasksContext,
when RemoveImage ends it tries to update the step of its parent job,
LiveMigrateDiskCommand.
The problem is that when RemoveImage takes a lot of time, for example
when the disk has WAD = true, the parent job ends before RemoveImage
ends, and the job entry is removed from the DB. Then, when RemoveImage
tries to update the job step, it fails with a violation of the job ID
foreign key constraint (because the job doesn't exist anymore).

This patch changes the call to RemoveImage to be done via a cloned and
detached context from the parent command to avoid this issue.

Change-Id: I909f94d430aaa9d8335cf8b6675e3db2ce0cb184
Bug-Url: https://bugzilla.redhat.com/1534241
Signed-off-by: Idan Shaby <ishaby@redhat.com>