core: Catch Exception type instead of XPathException
In order to improve logging when parsing OVF and avoid swallowed exceptions
, this patch changes the exception type which is catched
on XML handling, from XPathExpressionException to Exception.
That way, it is guerentee, that the string which is being parsed will be
presented in the log and that way it will be more easy to know the root
cause of the failure.

Change-Id: Ia1ac0c0e1db8d053684d9b7b425cfcd0899d36fc
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>