core: back-end support for sata disks
Initial support for having hard-disks set with SATA bus.
We currently support having CD-ROMs that are set with SATA
bus, when such configuration exists in osinfo, but we expect
SATA bus to be also set for hard-disks as a default selection
in q35 (rather than IDE that is the default today).

This patch adds this support on the back-end side. It cannot
be set though without proper support at the front-end (UI/
REST-API) part.

Change-Id: Icf0155a4b4833d9f71e1289d54dcebb33a9acbba
Signed-off-by: Arik Hadas <ahadas@redhat.com>