core: Call endAction() of all child commands in ImportVmCommand
ImportVmCommand.endActionOnAllImageGroups() duplicates essentially
VmCommand.endActionOnDisks() method, but limits it needlessly to calling
endAction() of CopyImageGroupCommand, which causes endAction() of other
child commands not to be called. In particular, endAction() of
AddVmLeaseCommand is not called, making VmDynamic not updated with VM
lease information (thanks Eyal Shenitzky for pointing this out).

Also, the most part of ImportVmCommand.endSuccessfully() may be replaced
by call to endSuccessfully() of the parent class (VmCommand).

Change-Id: Id3fdf0d76ccf36fe301f018932e1f64b0ce20955
Bug-Url: https://bugzilla.redhat.com/1547991
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>