core: Remove snapshot memory volume representation
The snapshot object contains a legacy field call memoryVolume which
contained a series of UUIDs representing the storage pool, storage
domain, image and volume IDs of the two disks of which the hibernation
data consists of.
At some point those volumes were represented as actual disks in the
system and two fields were added, one to represent the disk ID of the
memory dump and one the disk ID of the metadata.
Since we have the IDs of the disks, the memory volume has become
redundant and can be removed, also due to the facat that both disks can
reside in different storage domains so the data there might not even be
accurate.
This patch handles the removal of the memory volume while keeping all
the flows as they were for now.
The next phase needed will be to write in the OVF new data which will
give us the option to define the memory disks in different storage
domains.

Change-Id: I66df906ded602bc63dd8ff2c1c37589a0ea6e2ae
Signed-off-by: Tal Nisan <tnisan@redhat.com>