engine : Host stuck in "connecting" state
When the storage domain is down and the vdsm
is manually stopped and started, the status of
the host in engine is not updated until InitVdsOnUp
command completes.

InitVdsOnUpCommand takes about 2 minutes to complete
when the storage domain is down. This delays the
status update on engine side. Persisiting the host
status before invoking InitVdsOnUp resolves the issue.

Change-Id: Idc02307189b4bdf711a9c658d33467c7d2cf6a61
Bug-Url: https://bugzilla.redhat.com/1556971
Signed-off-by: Ravi Nori <rnori@redhat.com>
(cherry picked from commit ce7c7bb488d0532a3ba95f8295936862d0992a7c)