core: Use waiting locking for UpateNetworkCommand
If an external network is updated during the automatic synchronization
of all networks on the same provider, the updating should wait for the
lock until the synchronization is finished.

This absolves users of the REST-API , e.g. ovirt-system-tests,
to handle this situation.

The UpateNetworkCommand is change in an separate commit, because this
change will only be backported to branches which supports updating of
external networks.

Change-Id: I68c730e458717ebab5feba515ab6ff8e053578b3
Signed-off-by: Dominik Holler <dholler@redhat.com>