core: Replace generic unlockVm() logic in ImportVmCommand
Generic unlockVm() logic inherited from VmCommand cannot work in
ImportVmCommand, because getVm() method is overridden in
ImportVmCommandBase. In the case of importing as a new entity, after
calling setVm(null), the overridden getVm() still returns a VM object
from command parameters instead of refreshing it from the database. This
makes checking the VM status in this object useless.

Change-Id: I059c0d1301ea9aa2c2889c5e0a115d160a1def54
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>