aaa: Failed logging attempts are not audited / logged
Failed login attempts should be logged
so the events appear in engine's audit log.

Sso can use the client notification callback
to send audit log events to engine. The request
requires the client secret inorder to be able to
insert audit log events, so it is secure against
malicious attacks.

Change-Id: I8854bee7d20a1c662be36b9ee1661b7caed0cf27
Bug-Url: https://bugzilla.redhat.com/1550135
Signed-off-by: Ravi Nori <rnori@redhat.com>