core: OvfOvirtWriter log level
It is perfectly OK for a VM to have no users with permissions to it,
especially in smaller scale systems where a single admin creates and
managers VMs for users who don't use the self-service user portal.

This patch reduces logging such a situation from the WARN level to
DEBUG level.

Change-Id: I943ba13a8459ae82c773db9d7fd1ee3fcc560841
Bug-Url: https://bugzilla.redhat.com/1558924
Signed-off-by: Allon Mureinik <amureini@redhat.com>