webadmin: lease removed when VM edited via UI
VM lease will be removed while editing a VM via the UI
in case the VM lease storage domain is not active.
The VM appears to be without a lease and when the user press 'OK' the
lease removed.

This patch changes the behavior of the VM lease storage domain selection
box, if the lease storage domain isn't active the selection box will
show the non-active domain and it will be grayed out so the user
cannot edit it.

Change-Id: I19961cf8e625e823ddbfe6b654c73c3356790eeb
Bug-Url: https://bugzilla.redhat.com/1560455
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>