webadmin: Fix move disk dialog
Screenshot: https://imgur.com/a/Wvq8i

In case of trying to move a disk from an inactive storage domain,
the SD's name doesn't show, the target SD should be disabled and an
appropriate message saying "source storage domain is not active" should
appear as a tooltip.

In addition, in case of trying to copy a template disk which has more
than one source sd, the source sd name appears as a label instead of a
list box. Therefore, the user can't choose the source sd.

This patch changes all the above.

Change-Id: I4e76fc3ddc2e86ccae2dde2eeffb0fbc623924ba
Bug-Url: https://bugzilla.redhat.com/1551934
Signed-off-by: Shani Leviim <sleviim@redhat.com>