core: Migration duration from the first progress event
Calculate the actual VM migration duration from the moment when the
first migration progress event is received by the Engine. This will
exclude from the duration the period when the migration command is
queued in VDSM and not actually started. The total migration duration is
calculated as previously, from the start of the execution phase of
MigrateVmCommand.

Change-Id: I64408d537afbdc0621d40c73d40edb4f1c32de58
Bug-Url: https://bugzilla.redhat.com/1539777
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>