core: Add VM lease storage domain status validation before running a VM
Before this patch, it was allowed to run a VM with a lease on
a non-active storage domain.
The activation of the VM failed due to san-lock error (No space left).

This patch adds a validation for the VM lease storage domain status.
If the domain is not active the VM will not be able to run and a proper
message will present to the user.

Change-Id: I45a9c155fa68ffc7599870e2ebc9a51e593943f9
Bug-Url: https://bugzilla.redhat.com/1561006
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>