packaging: Add playbook to configure ovn on cluster
Adding playbook which allows to configure (or re-configure)
ovn controllers on an ovirt cluster.
This is needed to allow the user to easitly reconfigure a cluster
to use a network other than ovirtmgmt for ovn tunneling.
The playbook will allow the user to do this from the command
line by issuing the following command:
' ansible-playbook
    -i /usr/share/ovirt-engine-metrics/bin/ovirt-engine-hosts-ansible-inventory
    --extra-vars "
      cluster_name=<cluster name>
      ovn_central=<ovn central ip>
      ovn_tunneling_interface=<vdsm network name>"
'

Change-Id: I233f0b92d9f2589acf60f065374a5584ae3580c8
Signed-off-by: Marcin Mirecki <mmirecki@redhat.com>