build: Move OsinfoValidationsTest to utils
This patch moves the test of the default properties to the utils
project, which provides OsinfoPropertiesParser.

Since the parser and the format of the defaults file are coupled
anyway, it makes sense to test them together. Moreover, moving this
test to the utils module allows using the default tools and
configurations from the parent pom instead of reinventing the wheel
in a non-standard module outside of the normal maven hierarchy.

Change-Id: I5325703908f5bb13c5da99ca023224a6b2b1f97a
Signed-off-by: Allon Mureinik <amureini@redhat.com>